test_requirements.txt
mixer
pytz
django-libs
fabric3
mock
coverage
django-coverage
ipdb
flake8
tox
